refactoring bloc codes
