+++ lib/dependency_injector.dart
@@ -5,7 +5,7 @@ import 'package:word_translator/core/network/network_info.dart';
 import 'package:word_translator/features/translator_app/data/datasources/translator_remote_data_src.dart';
 import 'package:word_translator/features/translator_app/domain/repositories/translator_repository.dart';
 import 'package:word_translator/features/translator_app/domain/usecases/get_translated_text.dart';
-import 'package:word_translator/features/translator_app/presentation/bloc/translator_bloc.dart';
+import 'package:word_translator/features/translator_app/presentation/bloc/translator_bloc/translator_bloc.dart';
 
 import 'features/translator_app/data/repositories/translator_repository_impl.dart';
