+++ lib/features/translator_app/presentation/bloc/translator_bloc/translator_event.dart
