+++ lib/features/translator_app/presentation/widgets/search_container.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:word_translator/features/translator_app/presentation/bloc/translator_bloc.dart';
+import 'package:word_translator/features/translator_app/presentation/bloc/translator_bloc/translator_bloc.dart';
 
 class SearchContainer extends StatefulWidget {
   final String? from;
